analysis-service/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# AI / ML
transformers==4.36.0
torch>=2.2.0
google-generativeai==0.3.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Scheduler
apscheduler==3.10.4

# Service Discovery
py-eureka-client==0.11.2

# Logging
loguru==0.7.2

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6
